backend/requirements.txt
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
openai>=1.12.0
chromadb>=0.4.22
